+++ Interface/AddOns/Diminish_Options/widgets/slider.lua
@@ -35,6 +35,8 @@ function Widgets:CreateSlider(parent, text, tooltipText, minValue, maxValue, val
     slider.valueText = value
 
     slider.tooltipText = tooltipText
+    slider:SetScript("OnEnter", Widgets.OnEnter)
+    slider:SetScript("OnLeave", GameTooltip_Hide)
     slider.callbackFunc = func
     slider:SetScript("OnValueChanged", OnValueChanged)
     _G[name .. "Low"]:SetText("")