+++ Interface/AddOns/WeakAuras/AuraEnvironment.lua
@@ -32,7 +32,8 @@ end
 if WeakAuras.IsClassic() then
   local WA_GetUnitAuraBase = WA_GetUnitAura
   WA_GetUnitAura = function(unit, spell, filter)
-    local name, icon, count, debuffType, duration, expirationTime, source, isStealable, nameplateShowPersonal, spellId, canApplyAura, isBossDebuff, castByPlayer, nameplateShowAll, timeMod = WA_GetUnitAuraBase(unit, spell, filter)
+    local name, icon, count, debuffType, duration, expirationTime, source, isStealable, nameplateShowPersonal, spellId,
+          canApplyAura, isBossDebuff, castByPlayer, nameplateShowAll, timeMod = WA_GetUnitAuraBase(unit, spell, filter)
     if spellId then
       local durationNew, expirationTimeNew = LCD:GetAuraDurationByUnit(unit, spellId, source, name)
       if duration == 0 and durationNew then
@@ -40,7 +41,8 @@ if WeakAuras.IsClassic() then
           expirationTime = expirationTimeNew
       end
     end
-    return name, icon, count, debuffType, duration, expirationTime, source, isStealable, nameplateShowPersonal, spellId, canApplyAura, isBossDebuff, castByPlayer, nameplateShowAll, timeMod
+    return name, icon, count, debuffType, duration, expirationTime, source, isStealable, nameplateShowPersonal, spellId,
+           canApplyAura, isBossDebuff, castByPlayer, nameplateShowAll, timeMod
   end
 end
 
@@ -467,9 +469,12 @@ local FakeWeakAurasMixin = {
     -- Note these shouldn't exist in the WeakAuras namespace, but moving them takes a bit of effort,
     -- so for now just block them and clean them up later
     genericTriggerTypes = true,
-    newFeatureString = true,
     spellCache = true,
     StopMotion = true,
+    -- We block the loaded table, even though it doesn't exist anymore,
+    -- because some versions of ZT Tracker overwrote region:Collpase() and
+    -- checked for WeakAuras.loaded in there
+    loaded = true
   },
   override = {
     me = GetUnitName("player", true),
@@ -481,7 +486,8 @@ local FakeWeakAurasMixin = {
         Private.AuraWarnings.UpdateWarning(current_uid, "FakeWeakAurasGetData", "warning",
                   L["This aura calls GetData a lot, which is a slow function."])
       end
-      return CopyTable(WeakAuras.GetData(id))
+      local data = WeakAuras.GetData(id)
+      return data and CopyTable(data) or nil
     end,
     clones = MakeDeprecated(Private.clones, "clones",
                 L["Using WeakAuras.clones is deprecated. Use WeakAuras.GetRegion(id, cloneId) instead."]),
@@ -522,11 +528,12 @@ local exec_env_custom = setmetatable({},
     elseif k == "DebugPrint" then
       return DebugPrint
     elseif k == "C_Timer" then
-      return Private.AuraEnvironmentWrappedSystem.Get("C_Timer",
-                                current_aura_env.id, current_aura_env.cloneId)
+      return current_aura_env and Private.AuraEnvironmentWrappedSystem.Get("C_Timer",
+                                      current_aura_env.id, current_aura_env.cloneId)
+                              or C_Timer
     elseif blockedFunctions[k] then
       blocked(k)
-      return function() end
+      return function(_) end
     elseif blockedTables[k] then
       blocked(k)
       return {}
@@ -572,7 +579,7 @@ local exec_env_builtin = setmetatable({},
       return PrivateForBuiltIn
     elseif blockedFunctions[k] then
       blocked(k)
-      return function() end
+      return function(_) end
     elseif blockedTables[k] then
       blocked(k)
       return {}
@@ -596,16 +603,25 @@ function env_getglobal_builtin(k)
   return exec_env_builtin[k]
 end
 
+local function firstLine(string)
+  local lineBreak = string:find('\n', 1, true)
+  if lineBreak then
+    return string:sub(1, lineBreak - 1)
+  end
+  return string
+end
+
 local function CreateFunctionCache(exec_env)
   local cache = {}
   cache.Load = function(self, string)
     if self[string] then
       return self[string]
     else
-      local loadedFunction, errorString = loadstring(string)
+      local loadedFunction, errorString = loadstring(string, firstLine(string))
       if errorString then
         print(errorString)
       else
+        --- @cast loadedFunction -nil
         setfenv(loadedFunction, exec_env)
         local success, func = pcall(assert(loadedFunction))
         if success then