+++ Interface/AddOns/WeakAuras/Modernize.lua
@@ -1549,7 +1549,7 @@ function Private.Modernize(data)
 
   if data.internalVersion < 58 then
     -- convert key use for talent load condition from talent's index to spellId
-    if WeakAuras.IsDragonflight() then
+    if WeakAuras.IsRetail() then
       local function migrateTalent(load, specId, field)
         if load[field] and load[field].multi then
           local newData = {}
@@ -1567,6 +1567,92 @@ function Private.Modernize(data)
         migrateTalent(load, specId, "talent")
         migrateTalent(load, specId, "talent2")
         migrateTalent(load, specId, "talent3")
+      end
+    end
+  end
+
+  if data.internalVersion < 59 then
+    -- convert key use for talent known trigger from talent's index to spellId
+    if WeakAuras.IsRetail() then
+      local function migrateTalent(load, specId, field)
+        if load[field] and load[field].multi then
+          local newData = {}
+          for key, value in pairs(load[field].multi) do
+            if value ~= nil and Private.talentInfo[specId] and Private.talentInfo[specId][key] then
+              newData[Private.talentInfo[specId][key][2]] = value
+            end
+          end
+          load[field].multi = newData
+        end
+      end
+      for triggerId, triggerData in ipairs(data.triggers) do
+        if triggerData.trigger.type == "unit" and triggerData.trigger.event == "Talent Known" then
+          local classId
+          for i = 1, GetNumClasses() do
+            if select(2, GetClassInfo(i)) == triggerData.trigger.class then
+              classId = i
+            end
+          end
+          if classId and triggerData.trigger.spec then
+            local specId = GetSpecializationInfoForClassID(classId, triggerData.trigger.spec)
+            if specId then
+              migrateTalent(triggerData.trigger, specId, "talent")
+            end
+          end
+        end
+      end
+    end
+  end
+
+  if data.internalVersion < 60 then
+    -- convert texture rotation
+    if data.regionType == "texture" then
+      if data.rotate then
+        -- Full Rotate is enabled
+        data.legacyZoomOut = true
+      else
+        -- Discreete Rotation
+        data.rotation = data.discrete_rotation
+      end
+      data.discrete_rotation = nil
+    end
+  end
+
+  if data.internalVersion < 61 then
+    -- convert texture rotation
+    if data.regionType == "texture" then
+      if data.legacyZoomOut then
+        data.rotate = true
+      else
+        data.rotate = false
+        data.discrete_rotation = data.rotation
+      end
+      data.legacyZoomOut = nil
+    end
+  end
+
+  -- version 62 became 64 to fix a broken modernize
+
+  if data.internalVersion < 63 then
+    if data.regionType == "texture" then
+      local GetAtlasInfo = C_Texture and C_Texture.GetAtlasInfo or GetAtlasInfo
+      local function IsAtlas(input)
+        return type(input) == "string" and GetAtlasInfo(input) ~= nil
+      end
+
+      if not data.rotate or IsAtlas(data.texture) then
+        data.rotation = data.discrete_rotation
+      end
+    end
+  end
+
+  if data.internalVersion < 64 then
+    if data.regionType == "dynamicgroup" then
+      if data.sort == "custom" and type(data.sortOn) ~= "string" or data.sortOn == "" then
+        data.sortOn = "changed"
+      end
+      if data.grow == "CUSTOM" and type(data.growOn) ~= "string" then
+        data.growOn = "changed"
       end
     end
   end