+++ Interface/AddOns/WeakAurasTemplates/TriggerTemplatesDataWrath.lua
@@ -2,7 +2,8 @@ local AddonName, TemplatePrivate = ...
 local WeakAuras = WeakAuras
 if not WeakAuras.IsWrathClassic() then return end
 local L = WeakAuras.L
-local GetSpellInfo, tinsert, GetItemInfo, GetSpellDescription, C_Timer, Spell = GetSpellInfo, tinsert, GetItemInfo, GetSpellDescription, C_Timer, Spell
+local GetSpellInfo, tinsert, GetItemInfo, GetSpellDescription, C_Timer, Spell
+    = GetSpellInfo, tinsert, GetItemInfo, GetSpellDescription, C_Timer, Spell
 
 -- The templates tables are created on demand
 local templates =
@@ -1271,20 +1272,6 @@ itemInfoReceived:SetScript("OnEvent", function()
     end)
   end
 end);
-
-
--- Enrich Display templates with default values
-for regionType, regionData in pairs(TemplatePrivate.Private.regionOptions) do
-  if (regionData.templates) then
-    for _, item in ipairs(regionData.templates) do
-      for k, v in pairs(TemplatePrivate.Private.regionTypes[regionType].default) do
-        if (item.data[k] == nil) then
-          item.data[k] = v;
-        end
-      end
-    end
-  end
-end
 
 
 TemplatePrivate.triggerTemplates = templates